LOG-95 Setup token-based security

+++ local_run.sh
@@ -1,3 +1,9 @@
+# Check arguments
+if [ $# -ne 2 ]; then
+    echo "Usage: $0 <backend_host> <with_frontend_bool>"
+    exit 1
+fi
+
 # BACKEND (LOCALHOST)
 echo "Stopping and removing all Docker containers"
 docker rm -f $(docker ps -a -q)
@@ -14,6 +20,10 @@ echo "Booting up backend containers"
 docker compose up -d
 
 # FRONTEND (LOCALHOST SEM DOCKER)
+if [ $2 = "false" ]; then
+    echo "Skipped frontend"
+    exit 0
+fi
 cd ../frontend
-echo "REACT_APP_API_HOST = 'localhost'" > .env
+echo "REACT_APP_API_HOST = '$1'" > .env
 npm start
